mt_evtMenu: FUN_003760a0 fix uVar18 float-to-unsigned-int bug (same pattern as FUN_00381bf0, u32->f32) - massive impact since uVar18=FUN_0038a220(...) is used 100+ times in this large function; diff 20808->13333, object size 24660B->17556B. Residual is a different, also-substantial bug: many stack locations initialized from global string/data pointer constants use integer lw/sw where retail uses float lwc1/swc1 (pointer-stored-as-float-bitpattern struct fields) - needs dedicated follow-up investigation of the local struct layout

+++ src/Event/mt_evtMenu.c
@@ -6644,7 +6644,7 @@ void FUN_003760a0(int param_1,int param_2,u32 param_3,u16 *param_4,int param_5)
 
   int iVar17;
 
-  u32 uVar18;
+  f32 uVar18;
 
   u32 uStack_f0;
